Difficulty select (EASY/MEDIUM/HARD x100), 300 packed levels
- Boot asks SELECT MODE, then 100 levels of that mode (B goes back)
- 300 puzzles packed 52 bytes each (solution nibbles + givens mask,
  15.6KB): the old char format could not fit 300 levels in 32KB
- puzzles.c: puzzle_solution/puzzle_given unpackers; difficulty is the
  index range, no stored enum; level ids are uint16_t
- ui_diff screen; select shows per-mode title, simplified centered
  rows (marker, number, star), DONE x/100 of the current mode
- pause/win show the level number within the mode; win 'is_last' =
  mode completed
- Smoke test walks diff->select->game (128 checks), host tests verify
  the packed format (digits, mask match, valid solutions)

+++ src/board.h
@@ -29,8 +29,8 @@
 
 #include "types.h"
 
-/* Load level `level` (0-based): reset grid and mistakes. */
-void board_load(uint8_t level);
+/* Load level `level` (0-based, 0-299): reset grid and mistakes. */
+void board_load(uint16_t level);
 
 /* Value of cell `idx` (0-80, row by row): 0 = empty, 1-9 = digit. */
 uint8_t board_get(uint8_t idx);
